+++ src/entries/wvui.ts
@@ -1,8 +1,10 @@
 import WvuiButton from '@/components/button/Button.vue';
 import WvuiInput from '@/components/input/Input.vue';
+import WvuiTypeaheadSuggestion from '@/components/typeahead-suggestion/TypeaheadSuggestion.vue';
 
 export default {
 	version: VERSION,
 	WvuiButton,
-	WvuiInput
+	WvuiInput,
+	WvuiTypeaheadSuggestion
 };